refactor: déplacement script dans hydrate
